Move "run_tests" entrypoint to the module `wft4galaxy.app.runner`
